+++ logchange-core/src/main/java/dev/logchange/core/infrastructure/persistance/changelog/FileChangelogRepository.java
@@ -10,9 +10,14 @@ import dev.logchange.core.domain.config.model.Config;
 import dev.logchange.core.format.md.changelog.MDChangelog;
 import dev.logchange.core.format.release_date.ReleaseDate;
 import dev.logchange.core.format.yml.changelog.entry.YMLChangelogEntry;
+import dev.logchange.core.format.yml.changelog.entry.YMLChangelogEntryConfigException;
+import dev.logchange.core.format.yml.config.YMLChangelogException;
+import lombok.SneakyThrows;
 import lombok.extern.java.Log;
 import org.apache.maven.plugins.changes.model.ChangesDocument;
 import org.apache.maven.plugins.changes.model.io.xpp3.ChangesXpp3Writer;
+
+import com.fasterxml.jackson.databind.exc.UnrecognizedPropertyException;
 
 import java.io.*;
 import java.nio.charset.StandardCharsets;
@@ -151,11 +156,27 @@ public class FileChangelogRepository implements ChangelogRepository {
     }
 
     private List<ChangelogEntry> getEntries(File versionDirectory) {
-        return getEntriesFiles(versionDirectory)
+        List<Exception> exceptions = new ArrayList<>();
+
+        List<ChangelogEntry> entries = getEntriesFiles(versionDirectory)
                 .sequential()
-                .map(file -> YMLChangelogEntry.of(getEntryInputStream(file)))
+                .map((file) -> {
+                    try {
+                        return YMLChangelogEntry.of(getEntryInputStream(file), file.getPath());
+                    } catch (YMLChangelogEntryConfigException e) {
+                        exceptions.add(e);
+                        return null;
+                    }
+                })
+                .filter(Objects::nonNull)
                 .map(YMLChangelogEntry::to)
                 .collect(Collectors.toList());
+
+        if (!exceptions.isEmpty()) {
+            throw new YMLChangelogException(exceptions);
+        }
+
+        return entries;
     }
 
     private Stream<File> getEntriesFiles(File versionDirectory) {