+++ CHANGELOG.md
@@ -7,6 +7,66 @@
 <!-- !!! ⚠️ DO NOT MODIFY THIS FILE, YOUR CHANGES WILL BE LOST ⚠️ !!! -->
 
 
+[1.7.0] - 2024-09-13
+--------------------
+
+### Dependency updates (5 changes)
+
+- Upgraded org.apache.maven:maven-core from 3.9.8 to 3.9.9 (@marwin1991)
+- Upgraded org.apache.maven:maven-plugin-api from 3.9.6 to 3.9.9 (@marwin1991)
+- Upgraded org.apache.maven.plugins:maven-site-plugin from 3.12.1 to 3.20.0 (@marwin1991)
+- Upgraded org.apache.maven.plugins:maven-surefire-plugin from 3.3.1 to 3.5.0 (@marwin1991)
+- Upgraded org.pitest:pitest-maven from 1.16.1 to 1.16.3 (@marwin1991)
+
+
+[1.6.0] - 2024-09-10
+--------------------
+
+### Dependency updates (41 changes)
+
+- Upgraded com.fasterxml.jackson.core:jackson-databind from 2.17.0 to 2.17.1 (@marwin1991)
+- Upgraded com.fasterxml.jackson.core:jackson-databind from 2.17.1 to 2.17.2 (@marwin1991)
+- Upgraded com.fasterxml.jackson.dataformat:jackson-dataformat-yaml from 2.17.0 to 2.17.1 (@marwin1991)
+- Upgraded com.fasterxml.jackson.dataformat:jackson-dataformat-yaml from 2.17.1 to 2.17.2 (@marwin1991)
+- Upgraded commons-io:commons-io from 2.16.0 to 2.16.1 (@marwin1991)
+- Upgraded org.apache.commons:commons-text from 1.11.0 to 1.12.0 (@marwin1991)
+- Upgraded org.apache.maven:maven-artifact from 3.9.6 to 3.9.7 (@marwin1991)
+- Upgraded org.apache.maven:maven-artifact from 3.9.7 to 3.9.8 (@marwin1991)
+- Upgraded org.apache.maven:maven-artifact from 3.9.8 to 3.9.9 (@marwin1991)
+- Upgraded org.apache.maven:maven-core from 3.9.6 to 3.9.7 (@marwin1991)
+- Upgraded org.apache.maven:maven-core from 3.9.7 to 3.9.8 (@marwin1991)
+- Upgraded org.apache.maven:maven-core from 3.9.8 to 3.9.9 (@marwin1991)
+- Upgraded org.apache.maven:maven-plugin-api from 3.9.6 to 3.9.9 (@marwin1991)
+- Upgraded org.apache.maven.plugin-tools:maven-plugin-annotations from 3.12.0 to 3.13.0 (@marwin1991)
+- Upgraded org.apache.maven.plugin-tools:maven-plugin-annotations from 3.13.0 to 3.13.1 (@marwin1991)
+- Upgraded org.apache.maven.plugin-tools:maven-plugin-annotations from 3.13.1 to 3.15.0 (@marwin1991)
+- Upgraded org.apache.maven.plugins:maven-failsafe-plugin from 3.2.5 to 3.3.0 (@marwin1991)
+- Upgraded org.apache.maven.plugins:maven-failsafe-plugin from 3.3.0 to 3.3.1 (@marwin1991)
+- Upgraded org.apache.maven.plugins:maven-failsafe-plugin from 3.3.1 to 3.5.0 (@marwin1991)
+- Upgraded org.apache.maven.plugins:maven-gpg-plugin from 3.2.2 to 3.2.3 (@marwin1991)
+- Upgraded org.apache.maven.plugins:maven-gpg-plugin from 3.2.3 to 3.2.4 (@marwin1991)
+- Upgraded org.apache.maven.plugins:maven-gpg-plugin from 3.2.4 to 3.2.5 (@marwin1991)
+- Upgraded org.apache.maven.plugins:maven-javadoc-plugin from 3.7.0 to 3.10.0 (@marwin1991)
+- Upgraded org.apache.maven.plugins:maven-javadoc-plugin from 3.6.3 to 3.7.0 (@marwin1991)
+- Upgraded org.apache.maven.plugins:maven-plugin-plugin from 3.12.0 to 3.13.0 (@marwin1991)
+- Upgraded org.apache.maven.plugins:maven-plugin-plugin from 3.13.0 to 3.13.1 (@marwin1991)
+- Upgraded org.apache.maven.plugins:maven-plugin-plugin from 3.13.1 to 3.15.0 (@marwin1991)
+- Upgraded org.apache.maven.plugins:maven-surefire-plugin from 3.2.5 to 3.3.0 (@marwin1991)
+- Upgraded org.apache.maven.plugins:maven-surefire-plugin from 3.3.0 to 3.3.1 (@marwin1991)
+- Upgraded org.assertj:assertj-core from 3.25.3 to 3.26.0 (@marwin1991)
+- Upgraded org.assertj:assertj-core from 3.26.0 to 3.26.3 (@marwin1991)
+- Upgraded org.mockito:mockito-junit-jupiter from 5.11.0 to 5.12.0 (@marwin1991)
+- Upgraded org.pitest:pitest-maven from 1.15.8 to 1.16.0 (@marwin1991)
+- Upgraded org.pitest:pitest-maven from 1.16.0 to 1.16.1 (@marwin1991)
+- Upgraded org.projectlombok:lombok from 1.18.32 to 1.18.34 (@marwin1991)
+- Upgraded org.xmlunit:xmlunit-assertj3 from 2.9.1 to 2.10.0 (@marwin1991)
+- Upgraded renovatebot/github-action from v40.1.7 to v40.1.10 (@marwin1991)
+- Upgraded renovatebot/github-action from v40.1.10 to v40.1.11 (@marwin1991)
+- Upgraded renovatebot/github-action from v40.1.11 to v40.1.12 (@marwin1991)
+- Upgraded renovatebot/github-action from v40.1.12 to v40.2.2 (@marwin1991)
+- Upgraded renovatebot/github-action from v40.2.2 to v40.2.8 (@marwin1991)
+
+
 [1.5.0] - 2024-04-06
 --------------------